+++ merch-dropper/src/components/Shirt/ThumbDisplay.js
